add open-ended task
